Make provider/resource compilition fail early if interfaces aren't satisifed.
This can help because Go interfaces are implicit. We might incorrectly implement a method (wrong parameter types) and it will subsequentially never be called. To verify right away if we have all methods expected by the interfaces implemented as expected, we assign them explicitly and force the interface check early.