FIX: Model fields must be public.
To get `Plan.Get()` and `State.Get()` to work, the fields in the models we want to insert the data into must be public. This is because it internally uses the Go reflection API.

To make a struct field public, it's name must start with an uppercase letter.

The naming is only relevant for the code, because in terraform and the schema we always refer to the mapped name from the struct definition.