Feature: Allow overwriting the `Host` header for all request.
This is to work around situations where DNS isn't available (yet) but Gotify is behind a reverse-proxy which is routing traffic based on Hostnames. For these scenarios, we can simply use the static IP as the Endpoint and overwrite the `Host` header to what the reverse proxy requires to correctly route the request.