Removed `id` from `gotify_plugin`.
The `id` is set by the Gotify server, however it's unknown if it's really stable. Since we use the `module_path` to identify the plugin anyways (which is required to be unique), we can just use it. We'll make an extra request to find the ID of the plugin and _then_ make the API requests for that specific ID.

This reduces the state we need to take care of and removes some possible errors when the ID _does_ change (for some reason).