Use relative absolute for `dev_overrides`.
This breaks with relative paths. Terraform is fine with it, but the `tfplugindocs` command to generate docs breaks. We also don't want to hardcode an ENV that works on _my_ machine but none others.

So now we expand the path automatically via Direnv and refer to this ENV variable in the `dev_overrides`, which is possible since https://github.com/hashicorp/terraform/pull/34873 - the previous upgrade to Terraform 1.9.3 enables usage of this option.